% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableGenes.r
\name{availableGenes}
\alias{availableGenes}
\title{lists the available genes for all the presented marks}
\usage{
availableGenes(x)
}
\arguments{
\item{x}{a data list in Aidan's global_choice()$tmp format.
may have been prefiltered for marks and cell lines}
}
\value{
a vector of available genes
}
\description{
lists the available genes for all the presented marks
}
