% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableGenesByMarks.r
\name{availableGenesByMarks}
\alias{availableGenesByMarks}
\title{returns a list of all the available gene lists for each mark}
\usage{
availableGenesByMarks(x)
}
\arguments{
\item{x}{a list containing data for each mark type. The data entry
is assumed to be the $res component of each slot.}
}
\value{
a list. For each slot in the input list, names of fully
   represented genes in the slot $res component are listed
}
\description{
Designed to be used with \link{listIntersect}, this returns a
list of all the genes completely recorded (no missing cell lines)
as a list.
}
