% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colourLegendPanel.r
\name{colourLegendPanel}
\alias{colourLegendPanel}
\title{allows user to select colour legend mappings}
\usage{
colourLegendPanel(id, input, output, ui = T)
}
\arguments{
\item{id}{prefix for Shiny id's}

\item{input}{Shiny 'input' list}

\item{output}{Shiny 'output' list}

\item{ui}{if TRUE defines the UI component, FALSE not currently implemented}
}
\value{
standard Shiny value
}
\description{
allows user to select colour legend mappings
}
