% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colourLookup.r
\name{colourLookup}
\alias{colourLookup}
\title{assigns colours to text terms using a matched list}
\usage{
colourLookup(terms, dict, ignore.case = T, default.colour = "black")
}
\arguments{
\item{terms}{the list of terms to be coloured, e.g. labels on a 
plot}

\item{dict}{a list of vectors of single word terms to be 
searched for in terms. If the list is names, its names
will be taken as the colours, otherwise a list of artifical 
names of the form colour.X will be used.}

\item{ignore.case}{if true (default) case is ignored}

\item{default.colour}{("black") the colour or value assigned to all non-matched terms}
}
\value{
a vector of length terms coloured by the name of the 
  entry with the first hit in dict.
}
\description{
Compares a dictionary of search terms against a list of elements
in a string vector to assign colour - or other other coding - to 
each element. In particular, in turn, for each item, we cycle 
through the dictionary (a list of words as a space deliminated string) 
and return a colour taken from the name of the corresponding dictionary 
entry if a match is detected. If multiple hits are detected, the first is
returned. For an illustration, see the examples.
}
\details{
Typically, you might used to to assign named colours to labels on
the basis of a text match in their description, or to generate a 
factor of arbitary levels (i.e. not explicit factor names) to be 
used to assign colours (or other plotting attributes) automatically.
}
\examples{
# using a named dictionary

colourLookup(c("a big dog","lots of fish","a small cat","a cat-dog"),
            list(red="dog",
                 blue="cat"))

# or

colourLookup(c("a big Dog","lots of fish","a small cat","a cat-dog"),
             list(red="dog",
                  blue="cat"),
             default.colour = "none",
             ignore.case=F)
              
# and without names 

colourLookup(c("a big dog","a small cat","a cat-dog"),
            list("dog",
                 "cat"))
}
