% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpMarkerProfile.r
\name{cpMarkerProfile}
\alias{cpMarkerProfile}
\title{plots marker profile from a CP tensor factorisation}
\usage{
cpMarkerProfile(U, pcs = NULL, markers = NULL, sigma = 1)
}
\arguments{
\item{U}{matrix containing the marker tensor components}

\item{pcs}{the components to be plotted. If NULL (default), are a listed}

\item{markers}{the names of the markers. Correspond to the rows of U. If NULL, 
they are automatically labelled mark 1, mark 2, etc.}

\item{sigma}{a scalar multipler applied to the marker scores before plotting.
By default +1, values of -1 can be used to reverse their sign if needed}
}
\value{
a ggplot object of a facetted plot, one panel for each component.
}
\description{
A support graph showing the CP-factorisation profile for one or more component as 
as selected by pcs. This is still developmental: does the CP-facotorisation 
always give the same answers?
}
\examples{

cp1=cp(M,5)
cpMarkerProfile(cp1$U[[3]],c(2,3),marker=dimnames(M@data)[[3]])

}
