% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldToThreeTensor.r
\name{foldToThreeTensor}
\alias{foldToThreeTensor}
\title{folds a list of marks into a tensor format}
\usage{
foldToThreeTensor(x)
}
\arguments{
\item{x}{a list of marks similar to}
}
\value{
a rTensor object of the form input-rows x input-columns x marks
}
\description{
Assumes that each mark has identical dimensions and dimnames.
}
\examples{
m=matrix(1:4,nrow=2,dimnames=list(c("a","b"),c("x","y")))

x=list(one=list(res=m+100),
       two=list(res=m+200),
       three=list(res=m+300))

a=foldToThreeTensor(x)
a
a@data
}
