% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullCellLines.r
\name{fullCellLines}
\alias{fullCellLines}
\title{lists the cell lines containing a given number of entries
for all marks}
\usage{
fullCellLines(x, n = 1)
}
\arguments{
\item{x}{a list of data types in Aidan's global_choice()$tmp format.}

\item{n}{the minimum of required data points. Normally the
default}
}
\value{
a list of cell line names statisfying the required
  condition.
}
\description{
For each cell line, checks that at least n entries are present
for each mark type. Normally requiring at least one (the default)
this allows you to rule out samples which are missing for at least
one of the mark types.
}
