% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legendMappingsParser.r
\name{legendMappingsParser}
\alias{legendMappingsParser}
\title{parses a legend text to a list or vice versa}
\usage{
legendMappingsParser(x)
}
\arguments{
\item{x}{a text or list as described above}
}
\value{
a list or text, the opposite of x
}
\description{
If passed a legend list, converts it to a text string suitable
for display in a text box and if passed a text, converts back to
a string. The list output is designed to work with \link{colourLookup}.
You should look at this function for details.
}
\details{
The text format is written as key (to appear in the legend) = (an
equals sign)
a list of match terms used colourLookup. A separate entry appears
on each line. The left hand key can be multiple words. If the right
hand value is omitted, the value from the left is copied across. Finally,
if no text is supplied, a NULL value is returned.
}
\examples{
a=list("neut"="neutrophil nx",
       "macro"="macrophage",
       "mono"="monocyte",
       "eryth"="erythroblast",
       "ab"="alpha-beta",
       "mega"="megakaryocyte",
        "endo"="endo")

b=legendMappingsParser(a)
b
legendMappingsParser(b)
}
