% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listIntersect.r
\name{listIntersect}
\alias{listIntersect}
\title{returns the intersect of a list of sets}
\usage{
listIntersect(x)
}
\arguments{
\item{x}{a flat list containing the sets to be compared}
}
\description{
Extends the base intersect() function to more than two sets
by using a list notation. Note that the input argument must be
flat list; multiple arguments currently not supported.
}
\examples{
listIntersect(list(1:6,2:5,4:10))
}
