% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listMarksByLine.r
\name{listMarksByLine}
\alias{listMarksByLine}
\title{lists the number of marks by line}
\usage{
listMarksByLine(x)
}
\arguments{
\item{x}{data object, a copy of the contents of Aidan's
global_choice()$tmp}
}
\value{
data frame of cell line x marks giving the number of
   data points for each combination
}
\description{
Data integration methods require a full data cube. This function
returns a table of the number of entries by cell line and marks.
All values should be equal for any follow on data integration
methods to work.
}
