% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listMarksByProject.r
\name{listMarksByProject}
\alias{listMarksByProject}
\title{lists availablity of markers by project}
\usage{
listMarksByProject(x)
}
\arguments{
\item{x}{}
}
\description{
lists availablity of markers by project
}
