% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberToStringPercent.r
\name{numberToStringPercent}
\alias{numberToStringPercent}
\title{returns number strings like "100%"}
\usage{
numberToStringPercent(x)
}
\arguments{
\item{x}{an input number}
}
\value{
a string of the form "x%" where x is the given number
}
\description{
A simple wrapper for paste0 for rescalling sizes etc.
}
\examples{
numberToStringPercent(50)
}
