% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistribution.r
\name{plotDistribution}
\alias{plotDistribution}
\title{plots data distributions}
\usage{
plotDistribution(x, n = 1000)
}
\arguments{
\item{x}{a list of the form of Aidan's global_choice()$tmp}

\item{n}{the number of samples to pull for each mark to
create the distribution density plot}
}
\value{
a ggplot2 plotting object
}
\description{
Creates plots of data distributions by mark.
}
