% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensorCpAnalysisInspectorPanel.r
\name{tensorCpAnalysisInspectorPanel}
\alias{tensorCpAnalysisInspectorPanel}
\title{main CP analysis inspector panel}
\usage{
tensorCpAnalysisInspectorPanel(id, input, output, ui = T)
}
\arguments{
\item{id}{prefix for Shiny id's}

\item{input}{Shiny 'input' list}

\item{output}{Shiny 'output' list}

\item{ui}{if TRUE defines the UI component, FALSE not currently implemented}
}
\value{
standard Shiny value
}
\description{
main CP analysis inspector panel
}
