% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensorDataOptions.r
\name{tensorDataOptions}
\alias{tensorDataOptions}
\title{supporting function containing reference data values}
\usage{
tensorDataOptions(key = NULL)
}
\arguments{
\item{key}{a character to select the required entry, if NULL (default)
a list of available key values are returned.}
}
\value{
the selected data value
}
\description{
A function returning global data values
}
\examples{
tensorDataOptions("colour.legend.mappings")
}
