% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensorDataPanel.r
\name{tensorDataPanel}
\alias{tensorDataPanel}
\title{implements data panel for tensor functionality}
\usage{
tensorDataPanel(id, input, output, ui = T)
}
\arguments{
\item{id}{prefix for Shiny id's}

\item{input}{Shiny 'input' list}

\item{output}{Shiny 'output' list}

\item{ui}{if TRUE defines the UI component, FALSE not currently implemented}
}
\value{
standard Shiny value
}
\description{
implements data panel for tensor functionality
}
