% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensorDataTransPanel.r
\name{tensorDataTransPanel}
\alias{tensorDataTransPanel}
\title{implements data transformation panel for tensor functionality}
\usage{
tensorDataTransPanel(id, input, output, ui = T)
}
\arguments{
\item{id}{prefix for Shiny id's}

\item{input}{Shiny 'input' list}

\item{output}{Shiny 'output' list}

\item{ui}{if TRUE defines the UI component, FALSE not currently implemented}
}
\value{
standard Shiny value
}
\description{
implements data transformation panel for tensor functionality
}
