% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensorTab.r
\name{tensorTab}
\alias{tensorTab}
\title{top level wrapper for a tensor tab}
\usage{
tensorTab(input, output, ui = T)
}
\arguments{
\item{input}{Shiny 'input' list}

\item{output}{Shiny 'output' list}

\item{ui}{if TRUE defines the UI component, o/w the server}
}
\description{
top level wrapper for a tensor tab
}
