% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformData.r
\name{transformData}
\alias{transformData}
\title{transforms data prior to analysis}
\usage{
transformData(x, selection = "none", lambda = 0)
}
\arguments{
\item{x}{a data object of the form of Aidan's global_choice()$tmp}

\item{selection}{one of 'none', 'sqrt', 'log10' [for log10(x+1)] or 'softLog10' [currently not
implemented].}

\item{lambda}{a non-zero parameter for 'softLog10'  if selected. A
value of zero (default) will reduce to log10.}
}
\value{
a transformed copy of x
}
\description{
Supports a selection of simple data transformations to be
applied for prior to analysis. Note that this can still be
used after quantile normalisation: quantile normalisation
standardises samples to a common distribution but does not
its final shape. This is the function of a standard data
transformation.
}
