% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listMarksByProject.r
\name{listMarksByProject}
\alias{listMarksByProject}
\title{DEFUNCT lists availablity of markers by project}
\usage{
listMarksByProject(x)
}
\arguments{
\item{x}{a data object returned by Aidan's global_choice()}
}
\value{
a charactor vector with the names of the contained marks
}
\description{
Think this no longer works - just gives a list of marks ignorning projects
}
